work on devil fire projectile

+++ ZombiePaintball/Bullet.pde
@@ -1,12 +1,12 @@
 class Bullet {
-  int XCoord;
-  int YCoord;
-  int velX;
-  int velY;
-  int dmg;
-  int direction;
+  float XCoord;
+  float YCoord;
+  float velX;
+  float velY;
+  float dmg;
+  float direction;
 
-  Bullet(int X, int Y, int velX, int velY, int dmg, int dir) {
+  Bullet(float X, float Y, float velX, float velY, float dmg, float dir) {
     XCoord = X;
     YCoord = Y;
     this.velX = velX;
@@ -24,13 +24,13 @@ class Bullet {
     for(int i = 0; i < enemy.size; i++){
       if(dist(enemy.wave.get(i).XCoord,enemy.wave.get(i).YCoord,XCoord,YCoord) <= 15){
         knockback(enemy.wave.get(i),10);
-        enemy.wave.get(i).takeDamage(dmg);   
+        enemy.wave.get(i).takeDamage((int)dmg);   
         return true;
       }
     }
     return false;
   }
-  void knockback(Zombie x, int units){
+  void knockback(Zombie x, float units){
     if (direction == 0){
       x.YCoord -= units;
     }else if (direction == 2){
@@ -53,6 +53,7 @@ class Bullet {
       x.YCoord -= units;
     }
   }
+ 
  
   void move() {
     XCoord += velX;
